compress workers into one file
