+++ public/worker.js
@@ -1,15 +1,34 @@
-importScripts("./snarkjs.min.js");
-self.addEventListener("message", async (evt) => {
-  console.log("web worker recieved message");
-  const [input, zkeyFastFile] = evt.data;
-  console.log("🚀 ~ self.addEventListener ~ zkeyFastFile", zkeyFastFile);
-  console.log("🚀 ~ self.addEventListener ~ zkeyFastFile", typeof zkeyFastFile);
-  console.log("🚀 ~ self.addEventListener ~ input", input);
-  console.log("🚀 ~ self.addEventListener ~ input", typeof input);
-  const result = await snarkjs.groth16.fullProve(
-    input,
-    "/jwt.wasm",
-    zkeyFastFile
-  );
-  postMessage(result);
-});
+importScripts('./snarkjs.min.js')
+self.addEventListener('message', async ({ data }) => {
+  const fn = data[0]
+  switch (fn) {
+    case 'fullProve':
+      const [_, input, zkeyFastFile] = data
+      const result = await snarkjs.groth16.fullProve(
+        input,
+        '/jwt.wasm',
+        zkeyFastFile
+      )
+      postMessage(result)
+      break
+    case 'exportSolidityCallData':
+      const [__, proof, publicSignals] = data
+      const rawCallData = await snarkjs.groth16.exportSolidityCallData(
+        proof.data,
+        publicSignals.data
+      )
+      postMessage(rawCallData)
+      break
+    // case 'verify':
+    //   const [___, vkey, proof, publicSignals] = data
+    //   const proofVerified = await snarkjs.groth16.verify(
+    //     vkey,
+    //     publicSignals.data,
+    //     proof.data
+    //   )
+    //   postMessage(proofVerified)
+    //   break
+    default:
+      break
+  }
+})
